+++ lib/pure_result.dart
@@ -1,3 +1,4 @@
 export 'src/async_result.dart';
 export 'src/result.dart';
+export 'src/try_async_result.dart';
 export 'src/try_result.dart';